StatusType; js_only.
